requirements.txt
pandas==2.1.1
scikit-learn==1.3.1
matplotlib==3.8.0
chardet==5.2.0
openai==1.24.0
#openai==1.54.4
pandasql==0.7.3
streamlit==1.40.0
streamlit-antd-components==0.3.2
streamlit-browser-session-storage==0.0.3
streamlit-chat==0.1.1
streamlit-local-storage==0.0.12
streamlit-screen-stats==0.0.28
besser-bot-framework==1.5.0

Django==4.2.7
django-allauth==0.57.0
pyjwt[crypto]==2.8.0
pytest-cov==4.0.0
coverage[toml]==7.3.2
mkdocs==1.5.3
markdown==3.3.7
opencv-python==4.8.1.78
nltk==3.8.1
audio-recorder-streamlit
# audio-recorder-streamlit==0.0.8
# openapi-python-client==0.15.2
httpx==0.27.2
# typer==0.7.0
# spacy==3.5.4
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl


# besser-bot-framework @ git+https://github.com/BESSER-PEARL/BESSER-Bot-Framework.git@dev
